requirements.txt
# Core Dependencies
gymnasium>=0.29.0
numpy>=1.24.0
pygame>=2.5.0

# Computer Vision
opencv-python>=4.8.0
scipy>=1.11.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Reinforcement Learning
stable-baselines3>=2.1.0
sb3-contrib>=2.1.0

# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Data Processing
pandas>=2.0.0
